### 2025-11-24 #9506 @editorialbot
[PRE REVIEW]: solve_nivp: A Python toolkit for integrating nonsmooth dynamical systems
**Submitting author:** @dmriley993 (<a href="http://orcid.org/0000-0002-5463-5357">David Riley</a>)
**Repository:** https://github.com/ERC-INJECT/solve_nivp
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
