### 2025-11-24 #9506 @danielskatz
@dmriley993 - Your submission looks reasonable for us to review, but I don't see an appropriate editor currently available. I'll put this on our waiting list for now, but hopefully it won't be too long.

While this is in progress, you could work on the possibly missing and incorrect DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. An editorialbot command needs to be the first entry in a new comment.
