### 2025-12-02 866cbb6
fixed bug in order and improved sparsity
