+++ solve_nivp/nonlinear_solvers.py
@@ -1,5 +1,3 @@
-
-
 # nonlinear_solvers.py  (fast-path projector dispatch)
 
 from __future__ import annotations
@@ -33,7 +31,7 @@ class ImplicitEquationSolver:
         sparse: bool | str = 'auto',
         sparse_threshold: int = 200,
         linear_solver: str = 'gmres',
-        precond_reuse_steps: int = 5,
+        precond_reuse_steps: int = 20,
         ilu_drop_tol: float = 1e-4,
         ilu_fill_factor: float = 10.0,
         gmres_tol: float = 1e-6,
@@ -129,8 +127,22 @@ class ImplicitEquationSolver:
         self._last_shape = None
         self._last_pattern = None
 
+        # SPLU cache
+        self._lu = None
+        self._lu_use_count = 0
+        self._lu_pattern = None
+        self._lu_shape = None
+
         # Identity caches
         self._I_cache = {}
+
+        # Jacobian structure cache (for dense-to-sparse optimization)
+        self._J_cached = None
+        self._J_rows = None
+
+        # Tangent structure cache (for dense-to-sparse optimization)
+        self._D_cached = None
+        self._D_rows = None
 
         # Initialize rho state (scalar default + structured cache)
         self._set_initial_rho(rho0)
@@ -435,11 +447,8 @@ class ImplicitEquationSolver:
                 J_in = self._B
                 used_broyden = True
             else:
-                if self.jacobian is not None:
-                    J_in = self.jacobian(y)
-                else:
-                    mode = 'cs' if getattr(self, 'autodiff_mode', 'numerical') == 'cs' else 'fd'
-                    J_in = self._numerical_jacobian(func, y, sparse=sparse_active, mode=mode)
+                # Use optimized Jacobian computation with caching
+                J_in = self._compute_jacobian_csr(func, y, sparse_active)
 
             # Eisenstat–Walker tol for GMRES if enabled
             rtol_dyn = self.gmres_tol
@@ -448,21 +457,29 @@ class ImplicitEquationSolver:
                 rtol_dyn = max(self.gmres_tol, eta)
 
             if sparse_active:
-                # Use matrix-free linear operator to avoid forming (I - D) + lam * D @ J_in
+                # Sparse path: either use matrix-free GMRES (default) or explicit matrix + SPLU.
                 J_in = self._to_csr(J_in)
-                Dproj = self._to_csr(self._tangent(candidate, y, lam, tcur, F_in, prev), n)
+                Dproj = self._compute_tangent_csr(candidate, y, lam, tcur, F_in, prev, n)
                 rhs = -F_buf
 
-                # Define matvec and rmatvec for LinearOperator: J = I - D + lam * D @ J_in
-                def _matvec(v, _D=Dproj, _J=J_in, _lam=lam):
-                    return (v - _D @ v) + _lam * (_D @ (_J @ v))
+                if self.linear_solver == 'splu':
+                    # Form explicit sparse Jacobian: J = I - D + lam * D @ J_in
+                    J_mat = I - Dproj + lam * (Dproj @ J_in)
+                    if not sp.issparse(J_mat):
+                        J_mat = self._to_csr(J_mat, n)
+                    delta, ok = self._solve_linear_sparse(J_mat, rhs, rtol=rtol_dyn, pattern_hint=None)
+                else:
+                    # Matrix-free LinearOperator path: GMRES (with optional ILU preconditioning)
+                    def _matvec(v, _D=Dproj, _J=J_in, _lam=lam):
+                        return (v - _D @ v) + _lam * (_D @ (_J @ v))
 
[…]
