### 2025-12-04 c82f2bc
updated references
