+++ paper/biblio.bib
@@ -36,7 +36,7 @@
   address   = {London},
   isbn      = {978-1-84628-039-9},          
   doi       = {10.1007/978-1-84628-708-4},
-  url       = {https://doi.org/10.1007/978-1-84628-708-4},
+  url       = {doi.org/10.1007/978-1-84628-708-4},
   note      = {eBook ISBN 978-1-84628-708-4}
 }
 
@@ -152,7 +152,7 @@ number = {3},
 pages = {187-205},
 year = {2010},
 issn = {0167-2681},
-doi = {https://doi.org/10.1016/j.jebo.2010.02.008},
+doi = {10.1016/j.jebo.2010.02.008},
 author = {Fabio Tramontana and Frank Westerhoff and Laura Gardini},
 }
 
@@ -187,7 +187,8 @@ author = {Fabio Tramontana and Frank Westerhoff and Laura Gardini},
   publisher = {Stanford University Press},
   address   = {Stanford, CA},
   year      = {1958},
-  pages     = {154--165}
+  pages     = {154--165},
+  doi = {10.1017/cbo9780511664496.011},
 }
 
 @book{GlowinskiLionsTremolieres1981,
@@ -217,7 +218,7 @@ author = {Fabio Tramontana and Frank Westerhoff and Laura Gardini},
   pages={1--6},
   year={2012},
   publisher={Elsevier},
-doi = {https://doi.org/10.1016/j.compstruc.2012.01.009},
+doi = {10.1016/j.compstruc.2012.01.009},
 
 }
 
@@ -230,5 +231,5 @@ doi = {https://doi.org/10.1016/j.compstruc.2012.01.009},
   address={Berlin, Heidelberg},
   series={Springer Series in Computational Mathematics},
   volume={8},
-  doi={https://doi.org/10.1007/978-3-540-78862-1}
+  doi={10.1007/978-3-540-78862-1}
 }
