### 2025-12-09 #9506 @danielskatz
@jbytecode - sure, that's fine - I just saw that you finished one, so assumed that you might be able to take another.
