### 2026-01-17 #9775 @dmriley993
@danielskatz, we wanted to know whether the updated scope applies to our paper, given the pre-2026-submission tag?
