### 2026-01-17 #9775 @danielskatz
@dmriley993 - no, your submission will be judged under the old scope, and the reviewer checklists above have been updated to the old checklists.
