### 2026-01-29 #9775 @trintlermint
@danielskatz Hi! Everythings fine, my entire semester's exams end today actually, so I will go through the entire checklist on Friday/Saturday. So far I didnt have issues installing the package. =)
