### 2026-01-30 ERC-INJECT/solve_nivp#1 @ductho-le
Test failure: test_identity_tangent_is_identity_any_y with NumPy 2.x
## Description

During JOSS review, one test fails when running with NumPy 2.4.1 and SciPy 1.17.0:

```
FAILED tests/test_projections_added.py::test_identity_tangent_is_identity_any_y
NotImplementedError: subtracting a nonzero scalar from a sparse matrix
```

## Environment
- Python 3.11
- NumPy 2.4.1
- SciPy 1.17.0
- Windows 11

## Suggested Fix
Either:
1. Fix the sparse matrix operation to be compatible with NumPy 2.x
2. Or document/pin the supported NumPy version range in `pyproject.toml`

## JOSS Review
This is a blocking issue for the JOSS review checklist item "Functionality: Have the functional claims of the software been confirmed?"

Related: openjournals/joss-reviews#9775
