### 2026-01-30 ERC-INJECT/solve_nivp#2 @ductho-le
Add GitHub Actions workflow for automated testing
## Description

The repository currently only has a GitHub Actions workflow for generating the paper PDF (`draft-pdf.yml`). There is no automated testing workflow.

## Suggestion

Add `.github/workflows/tests.yml` to run pytest on multiple Python versions. Example:

```yaml
name: Tests

on:
  push:
    branches: [main]
  pull_request:
    branches: [main]

jobs:
  test:
    runs-on: ubuntu-latest
    strategy:
      matrix:
        python-version: ['3.9', '3.10', '3.11', '3.12']

    steps:
    - uses: actions/checkout@v4
    - name: Set up Python ${{ matrix.python-version }}
      uses: actions/setup-python@v5
      with:
        python-version: ${{ matrix.python-version }}
    - name: Install dependencies
      run: |
        python -m pip install --upgrade pip
        pip install -e .[test]
    - name: Run tests
      run: pytest -v
```

## JOSS Review
This addresses the JOSS review checklist item "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"

Related: openjournals/joss-reviews#9775
