### 2026-01-30 #9775 @ductho-le
I've begun my review and created the following issues for the authors to address:

- ERC-INJECT/solve_nivp#1 - Test failure with NumPy 2.x compatibility
- ERC-INJECT/solve_nivp#2 - Add GitHub Actions workflow for automated testing

Minor items (no separate issues):
- `CITATION.cff` contains placeholder ORCIDs and incorrect repository URL
- Debug `print(p)` statement in `solve_ivp_ns()` should be removed

I will complete my checklist once these are addressed.
