### 2026-02-02 43f6bf0
Add auto tests, fix sparse matrix test for NumPy 2.x, remove debug print
