### 2026-02-02 #9775 @dmriley993
@ductho-le I have responded to the two issues raised with more detail in the open issues. A quick summary of the response is provided here.

- For the first issue, the test failure was rectified by first converting to a dense matrix prior to comparison.

- I have added the actions workflow for automated testing

I have also updated the CITATION.cff for the ORCIDs and the correct repository URL as well as removing the the debug printing.
