### 2026-02-06 ERC-INJECT/solve_nivp#3 @wwhenxuan
Docs failure: intersphinx_mapping wrong in make html for cross-references
## Description

During the JOSS review process, I discovered an error in the compilation of the documentation. The specific problem lay in the cross-references of other related libraries.

~~~python
# Intersphinx mappings for cross-references
intersphinx_mapping = {
    'python': ('https://docs.python.org/3', {}),
    'numpy': ('https://numpy.org/doc/stable/', {}),
    'scipy': ('https://docs.scipy.org/doc/scipy/', {}),
}
~~~

## Environment

I updated the version of Sphinx and the problem persists.

* Platform:         win32; (Windows-10-10.0.22631-SP0)
* Python version:   3.11.10 (CPython)
* Sphinx version:   9.0.4
* Docutils version: 0.22.4
* Jinja2 version:   3.1.6
* Pygments version: 2.19.2

## Suggested Fix

According to the Sphinx documentation, the invalid list position value here cannot be an empty dictionary; it can be replaced with a None value.

~~~python
# Intersphinx mappings for cross-references
intersphinx_mapping = {
    'python': ('https://docs.python.org/3', None),
    'numpy': ('https://numpy.org/doc/stable/', None),
    'scipy': ('https://docs.scipy.org/doc/scipy/', None),
}
~~~

After making this modification, the docs can be compiled normally. By the way, it is beautifully designed.

## JOSS Review

This issue falls under the Functionality documentation section of the JOSS checklist.

Related: https://github.com/openjournals/joss-reviews/issues/9775
