### 2026-02-12 #9775 @danielskatz
@dmriley993 - Can you take a look at the [issue](https://github.com/ERC-INJECT/solve_nivp/issues/3) from @wwhenxuan above?
