### 2026-02-12 610ee24
fix: use None instead of {} in intersphinx_mapping (fixes #3)
