### 2026-02-12 ERC-INJECT/solve_nivp#3 @dmriley993
Thank you for the suggestion and comment. The fix has been implemented in commit [610ee24](https://github.com/ERC-INJECT/solve_nivp/commit/610ee24). The intersphinx_mapping values have been changed from {} to None as you recommended.
