### 2026-02-13 8d108a4
feat: SDIRK2 integrator, embedded adaptive bypass, cross-call SPLU reuse
- SDIRK2 (Alexander 1977, γ=1−√2/2): L-stable, 2nd-order, 2-stage DIRK
  with embedded 1st-order error estimate for adaptive stepping.
- Mass-matrix support: stage-2 shift and error use stage values directly
  (y_shift = y + ((1−γ)/γ)(Y1−y)), correct for general A ≠ I.
- AdaptiveStepping: embedded error bypass for methods with
  has_embedded_error=True — 1 integrator call vs 3 for Richardson.
- Cross-call SPLU factorisation reuse in _solve_newton_identity:
  self._lu persists across consecutive solve() calls so SDIRK2 sta
[…]
