+++ solve_nivp/adaptive_integrator.py
@@ -255,6 +255,59 @@ class AdaptiveStepping:
         # print(f"error is: {math.sqrt(accum / count) if count>0 else 0.0}")
         return 0.0 if count == 0 else math.sqrt(accum / count)
 
+    def _scaled_error_embedded(self, y_new, err_vec) -> float:
+        """Compute normalised RMS error from an embedded error vector.
+
+        Unlike :meth:`_scaled_error` (which builds the raw error from two
+        Richardson solutions), this method takes the element-wise error
+        estimate directly from the integrator (e.g. SDIRK2) and scales it
+        by the same tolerance formula::
+
+            tol_i = atol + rtol * |y_new_i|
+            E     = rms( err_i / tol_i )
+
+        Returns ``E`` with the same semantics: ``E <= 1`` means acceptable.
+        """
+        accum = 0.0
+        count = 0
+
+        if self.component_slices is None:
+            if self._err_buf is None or self._err_buf.shape != y_new.shape:
+                self._err_buf = np.empty_like(y_new)
+            if self._etol_buf is None or self._etol_buf.shape != y_new.shape:
+                self._etol_buf = np.empty_like(y_new)
+
+            # tol scaling: atol + rtol * |y_new|
+            np.abs(y_new, out=self._etol_buf)
+            self._etol_buf *= self.rtol
+            self._etol_buf += self.atol
+
+            # scaled error
+            np.divide(err_vec, self._etol_buf, out=self._err_buf)
+            accum = float(np.dot(self._err_buf.ravel(), self._err_buf.ravel()))
+            count = self._err_buf.size
+        else:
+            for i, sl in enumerate(self.component_slices):
+                if i in self.skip_error_indices:
+                    continue
+                yn_blk = y_new[sl]
+                er_blk = err_vec[sl]
+
+                if self._err_buf is None or self._err_buf.shape != yn_blk.shape:
+                    self._err_buf = np.empty_like(yn_blk)
+                if self._etol_buf is None or self._etol_buf.shape != yn_blk.shape:
+                    self._etol_buf = np.empty_like(yn_blk)
+
+                np.abs(yn_blk, out=self._etol_buf)
+                self._etol_buf *= self.rtol
+                self._etol_buf += self.atol
+
+                np.divide(er_blk, self._etol_buf, out=self._err_buf)
+                accum += float(np.dot(self._err_buf.ravel(), self._err_buf.ravel()))
+                count += self._err_buf.size
+
+        return 0.0 if count == 0 else math.sqrt(accum / count)
+
     # ------------------------------------------------------------------
     # Attempt logging helpers (optional)
     # ------------------------------------------------------------------
@@ -453,6 +506,12 @@ class AdaptiveStepping:
         """
         Attempt one adaptive step of size h starting from (t,y).
 
+        If the integrator exposes ``has_embedded_error = True``, a single
+        call to ``integrator.step()`` is made and the element-wise error
+        vector it returns is used directly for step-size control—no
+        Richardson extrapolation (step-doubling) is performed.  This cuts
+        the number of implicit solves from 3× to 1× per attempt.
+
         Returns
         -------
         (y_new, fk_new, h_next, E_curr, success, solver_error, iterations)
@@ -465,6 +524,104 @@ class AdaptiveStepping:
           - E_curr is the normalized local error estimate for this attempt.
           - success is True if we accept and advance, else False.
         """
+
+        # ==============================================================
+        # Fast path: integrator provides an embedded error estimate
+        # (e.g. SDIRK2).  Only ONE call to integrator.step() needed.
+        # ==============================================================
+        if getattr(self.integrator, 'has_embedded_error', False):
+            return self._step_embedded(fun, t, y, h)
+
[…]
