+++ solve_nivp/projections.py
@@ -258,37 +258,111 @@ class CoulombProjection(Projection):
             else:
                 self.constraint_indices = np.array([])
         self.use_numba = use_numba if use_numba in ('auto', True, False) else 'auto'
+        # Cache the call signature of con_force_func to avoid try/except on every call
+        self._con_force_nargs = None  # will be set on first call
 
     def _call_con_force(self, y, t=None, Fk_val=None):
-        # Flexible call to support signatures: (y), (y,t), (y,t,Fk_val)
-        try:
+        # On first call, determine the number of args accepted and cache it
+        nargs = self._con_force_nargs
+        if nargs == 3:
             return self.con_force_func(y, t, Fk_val)
+        elif nargs == 2:
+            return self.con_force_func(y, t)
+        elif nargs == 1:
+            return self.con_force_func(y)
+        # First call: probe signature
+        try:
+            result = self.con_force_func(y, t, Fk_val)
+            self._con_force_nargs = 3
+            return result
         except TypeError:
             try:
-                return self.con_force_func(y, t)
+                result = self.con_force_func(y, t)
+                self._con_force_nargs = 2
+                return result
             except TypeError:
-                return self.con_force_func(y)
+                result = self.con_force_func(y)
+                self._con_force_nargs = 1
+                return result
 
-    def _compute_jacobian(self, y, t=None, Fk_val=None):
+    def _compute_jacobian(self, y, t=None, Fk_val=None, rows=None):
+        """Compute the constraint force Jacobian.
+        
+        Parameters
+        ----------
+        rows : array-like of int, optional
+            If provided, only compute these rows of the Jacobian. Returns an
+            (len(rows), n) array instead of the full (n, n) Jacobian.
+        """
         if self.jac_func is not None:
             try:
                 J = self.jac_func(y, t, Fk_val)
             except TypeError:
                 J = self.jac_func(y)
-            return _to_numpy_if_needed(J)
-        return self._numerical_jacobian(y, t=t, Fk_val=Fk_val)
+            J = _to_numpy_if_needed(J)
+            if rows is not None:
+                rows = np.asarray(rows)
+                if sp.issparse(J):
+                    return J[rows].toarray() if hasattr(J[rows], 'toarray') else np.asarray(J[rows])
+                return J[rows]
+            return J
+        return self._numerical_jacobian(y, t=t, Fk_val=Fk_val, rows=rows)
 
-    def _numerical_jacobian(self, y, eps=1e-8, t=None, Fk_val=None):
+    def _numerical_jacobian(self, y, eps=1e-8, t=None, Fk_val=None, rows=None):
+        """Compute the constraint force Jacobian via finite differences.
+        
+        Parameters
+        ----------
+        rows : array-like of int, optional
+            If provided, only store these rows. Returns (len(rows), n) array.
+            When the constraint force Jacobian is diagonal (local constraint
+            force), only ``len(rows)`` perturbations are needed instead of n.
+        """
         y_np = _to_numpy_if_needed(y)
         n = len(y_np)
-        J = np.zeros((n,n), dtype=y_np.dtype)
         f0 = _to_numpy_if_needed(self._call_con_force(y, t=t, Fk_val=Fk_val))
-        for j in range(n):
-            y_pert = y_np.copy()
-            y_pert[j] += eps
-            f_eps = _to_numpy_if_needed(self._call_con_force(y_pert, t=t, Fk_val=Fk_val))
-            J[:, j] = (f_eps - f0)/eps
-        return J
[…]
