+++ docs/source/api/solve_nivp.ODESolver.rst
@@ -1,0 +1,23 @@
+solve\_nivp.ODESolver
+=====================
+
+.. currentmodule:: solve_nivp
+
+.. autoclass:: ODESolver
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~ODESolver.__init__
+      ~ODESolver.solve
+   
+   
+
+   
+   
+   
