+++ docs/source/api/solve_nivp.integrations.rst
@@ -1,0 +1,18 @@
+solve\_nivp.integrations
+========================
+
+.. automodule:: solve_nivp.integrations
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      AlgebraicBackwardEuler
+      BackwardEuler
+      CompositeMethod
+      EmbeddedBETR
+      IntegrationMethod
+      ThetaMethod
+      Trapezoidal
+   
