+++ docs/source/api/solve_nivp.rl.rst
@@ -1,0 +1,15 @@
+solve\_nivp.rl
+==============
+
+.. automodule:: solve_nivp.rl
+
+   
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   callbacks
+   dependency
+   env
