+++ docs/source/api/solve_nivp.ODESystem.rst
@@ -1,0 +1,25 @@
+solve\_nivp.ODESystem
+=====================
+
+.. currentmodule:: solve_nivp
+
+.. autoclass:: ODESystem
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~ODESystem.__init__
+      ~ODESystem.step
+      ~ODESystem.step_adaptive
+      ~ODESystem.step_fixed
+   
+   
+
+   
+   
+   
