+++ docs/source/api/solve_nivp.rst
@@ -1,0 +1,25 @@
+﻿solve\_nivp
+===========
+
+.. automodule:: solve_nivp
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      solve_ivp_ns
+   
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   ODESolver
+   ODESystem
+   adaptive_integrator
+   integrations
+   nonlinear_solvers
+   projections
+   rl
