+++ docs/source/api/solve_nivp.projections.rst
@@ -1,0 +1,17 @@
+solve\_nivp.projections
+=======================
+
+.. automodule:: solve_nivp.projections
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      CoulombProjection
+      GeneralMoreauVIProjection
+      IdentityProjection
+      MuScaledSOCProjection
+      Projection
+      SignProjection
+   

[…]
