### 2026-03-01 #9775 @wwhenxuan
Thank @dmriley993 for responding to my comments.

I have completed the checks in the review checklist, and all the examples, documentation, automated tests, and paper writing meet the requirements.
Thank you for your excellent open-source contributions to the field of nonsmooth dynamical systems.
