### 2026-03-03 c42780c
SDIRK2 SOC contact: block-diagonal rho, linesearch, contact module, tests
Phase 0 implementation for SDIRK2 + SOC frictional contact:

SDIRK2 fixes (integrations.py):
- step_size_arg = gamma*h (was h) so impulse coupling B*r/(theta*h_val)
  produces correct net impulse when stage eqn multiplies by gamma*h
- Block-diagonal rho equilibration: rho_i = gamma*h/A_ii (physical),
  rho_i = 1 (algebraic). Equilibrates natural-map Jacobian diagonal
  from ~3414 to O(1), fixing Newton conditioning for SDIRK2

Nonlinear solver (nonlinear_solvers.py):
- Vector lam support in 5 Jacobian assembly paths (spa
[…]
