+++ solve_nivp/adaptive_integrator.py
@@ -59,7 +59,7 @@ class AdaptiveStepping:
         component_slices=None,
         atol: float = 1e-6,
         rtol: float = 1e-3,
-        h0: float = 1e-2,
+        h0: float | str | None = 1e-2,
         h_min: float = 1e-10,
         h_max: float = 1e3,
         h_up: float = 2.0,
@@ -79,19 +79,42 @@ class AdaptiveStepping:
         verbose: bool = False,
         skip_error_indices=None,
         record_attempts: bool = False,
+        # --- DAE-aware error weighting ---
+        dae_var_weight: str = "auto",   # "auto", "exclude", "include"
     ) -> None:
 
         self.integrator = integrator
         self.component_slices = component_slices
-        self.atol = float(atol)
-        self.rtol = float(rtol)
 
-        # current step size h_n
-        self.h = float(h0)
+        # ---- Per-DOF tolerance vectors (SUNDIALS / scipy convention) ----
+        # Accept scalar, per-slice list/tuple, or full per-DOF array.
+        # Internally we store 1-D numpy arrays (or a scalar float when the
+        # system size is not yet known so that the first call lazily expands).
+        self._atol_raw = atol
+        self._rtol_raw = rtol
+        self._atol_vec: np.ndarray | None = None   # lazily built on first use
+        self._rtol_vec: np.ndarray | None = None
+
+        # ---- Automatic h₀ (Hairer-Wanner) ----
+        # When h0 is None or "auto", defer estimation to the first call to
+        # step(), where we have access to fun, t₀, y₀.
+        if h0 is None or (isinstance(h0, str) and h0.lower() == 'auto'):
+            self._auto_h0 = True
+            self.h = float(h_min)     # placeholder until estimated
+        else:
+            self._auto_h0 = False
+            self.h = float(h0)
 
         # global hard clamps on h
         self.h_min = float(h_min)
         self.h_max = float(h_max)
+
+        # ---- DAE-aware error weighting ----
+        # "auto"   : detect algebraic DOFs from mass matrix on first use
+        # "exclude": same as auto (kept for clarity)
+        # "include": traditional behaviour, weight all DOFs equally
+        self._dae_var_weight_mode = dae_var_weight.lower().strip()
+        self._dae_mask: np.ndarray | None = None    # lazily built
 
         # per-step up/down clamp factors (classic mode)
         self.h_up = float(h_up)
@@ -139,11 +162,232 @@ class AdaptiveStepping:
         self._attempt_status = None
         self.reset_attempt_log()
 
+        # --- Nonlinear failure recovery tracking ---
+        # Detects persistent fail→succeed cycles (the "death spiral" where
+        # every proposed h triggers a NL failure, the retry at h*h_down
+        # succeeds with tiny error, the controller grows h, and the next
+        # attempt fails again).
+        self._nl_fail_recovery_count = 0   # consecutive fail→succeed pairs
+        self._nl_success_no_fail = 0       # consecutive clean successes
+        self._in_nl_recovery = False        # previous step() had NL failure
+        self._consecutive_nl_fails = 0     # consecutive NL failures (no success between)
+
+        # How many consecutive NL failures before a full solver reset
+        # (destroy PETSc KSP, SPLU, ILU, all caches — force fresh Jacobian
+        # and fresh factorisation on the next attempt).
+        self._NL_RESCUE_THRESH = 5
+
         # operating mode
         m = mode.lower().strip()
         if m not in ("classic", "ratio"):
             raise ValueError("mode must be 'classic' or 'ratio'")
[…]
