+++ solve_nivp/projections.py
@@ -14,6 +14,8 @@ The norm of the projection residual ``||y - project(y, y - lam F(y))||`` drives
 nonlinear convergence tests in both VI and semismooth Newton solvers.
 """
 
+import inspect
+
 import numpy as np
 import scipy.sparse as sp
 from abc import ABC, abstractmethod
@@ -106,6 +108,535 @@ class IdentityProjection(Projection):
 
     def project_batch(self, current_state, candidates, rhok=None, t=None, Fk_val=None):
         return np.asarray(candidates)
+
+
+##############################################################################
+# _ConstraintBlock  (internal helper for AlgebraicConstraintProjection)
+##############################################################################
+def _count_required_args(fn):
+    """Count the number of **required** positional parameters of *fn*.
+
+    Parameters that have default values (common closure-capture pattern,
+    e.g. ``lambda p, _C=matrix: _C @ p``) are *not* counted.
+    Returns ``None`` if introspection fails (e.g. C built-ins).
+    """
+    if fn is None:
+        return 0
+    try:
+        sig = inspect.signature(fn)
+        return sum(
+            1 for p in sig.parameters.values()
+            if p.default is inspect.Parameter.empty
+            and p.kind in (inspect.Parameter.POSITIONAL_ONLY,
+                           inspect.Parameter.POSITIONAL_OR_KEYWORD)
+        )
+    except (ValueError, TypeError):
+        return None          # fall back to runtime probing
+
+
+class _ConstraintBlock:
+    """Per-constraint callable with arity-dispatch cache.
+
+    Each block stores one algebraic relation ``q = g(y)`` together with
+    its analytical Jacobian (or finite-difference fallback) and the
+    index slices that locate ``y`` and ``q`` inside the full state
+    vector.
+
+    Arity is detected at construction time via ``inspect.signature`` so
+    that user lambdas with closure-captured defaults (e.g.
+    ``lambda p, _C=matrix: _C @ p``) are correctly identified as
+    single-argument functions.
+    """
+    __slots__ = ('g', 'dg_dy', 'y_slice', 'q_slice', 'fd_eps',
+                 '_g_nargs', '_dg_nargs')
+
+    def __init__(self, g, dg_dy, y_slice, q_slice, fd_eps=1e-8):
+        self.g = g
+        self.dg_dy = dg_dy
+        self.y_slice = y_slice
+        self.q_slice = q_slice
+        self.fd_eps = float(fd_eps)
+        self._g_nargs = _count_required_args(g)
+        self._dg_nargs = _count_required_args(dg_dy)
+
+    # ---- flexible calling helpers ----
+    def call_g(self, y_sub, t=None, Fk_val=None):
+        """Evaluate ``g(y_sub)`` with auto-detected arity."""
+        if self.g is None:
+            s = self.q_slice
+            start = s.start or 0
+            stop = s.stop
+            step = s.step or 1
+            n_q = max(0, (stop - start + step - 1) // step) if stop is not None else y_sub.size
+            return np.zeros(n_q, dtype=float)
+        n = self._g_nargs
+        if n is not None:
+            if n >= 3:
+                return np.asarray(self.g(y_sub, t, Fk_val), dtype=float)
+            elif n == 2:
+                return np.asarray(self.g(y_sub, t), dtype=float)
+            else:
+                return np.asarray(self.g(y_sub), dtype=float)
+        # Fallback: runtime probing (only when inspect.signature failed)
+        try:
+            r = self.g(y_sub, t, Fk_val); self._g_nargs = 3
+            return np.asarray(r, dtype=float)
+        except TypeError:
+            try:
[…]
