### 2026-03-04 1cd3a52
feat: auto-generate component_slices from contact spec + active-set error filter
