+++ solve_nivp/projections.py
@@ -70,6 +70,37 @@ class Projection(ABC):
     def tangent_cone(self, candidate, current_state, rhok=None, t=None, Fk_val=None):
         pass
 
+    # ------------------------------------------------------------------
+    # Active-set filter helpers (overridden in SOC-based projections)
+    # ------------------------------------------------------------------
+    def regime_snapshot(self):
+        """Return an opaque snapshot of the per-block contact regime.
+
+        Returns *None* when the projection does not track regimes.
+        Subclasses (e.g. :class:`MuScaledSOCProjection`) override this.
+        """
+        return None
+
+    def regime_changed_mask(self, prev_snapshot, n_dof):
+        """Return a 1-D float mask of shape ``(n_dof,)`` for error weighting.
+
+        Entries are ``1.0`` (keep) for DOFs unaffected by a regime
+        transition, and ``0.0`` (suppress) for DOFs whose associated
+        contact block changed regime between *prev_snapshot* and the
+        current state.
+
+        Parameters
+        ----------
+        prev_snapshot : object
+            Opaque value returned by a previous :meth:`regime_snapshot`.
+        n_dof : int
+            Total dimension of the state vector.
+
+        Returns *None* when the projection does not track regimes or
+        *prev_snapshot* is *None*.
+        """
+        return None
+
     # Optional batched APIs (default fallbacks preserve behavior)
     def project_batch(self, current_state, candidates, rhok=None, t=None, Fk_val=None):
         candidates = np.asarray(candidates)
@@ -1869,6 +1900,115 @@ class MuScaledSOCProjection(Projection):
             if m > 0:
                 self._batch_branch_w_hat[:, 0] = 1.0
 
+    def _sync_branch_dict_to_batch(self):
+        """Copy per-block regime codes from ``_branch_region`` (dict)
+        into ``_batch_branch_region`` (ndarray).
+
+        The dense / sparse ``tangent_cone`` paths update the dict via
+        ``_proj_persistent``.  The batch fast-path updates the array
+        directly.  This helper keeps them in sync so that
+        ``regime_snapshot()`` always reads the latest data.
+        """
+        if self._uniform_m is None:
+            return
+        for k, region in self._branch_region.items():
+            if isinstance(k, int) and 0 <= k < len(self._batch_branch_region):
+                self._batch_branch_region[k] = region
+
+    # ------------------------------------------------------------------
+    # Active-set filter helpers (regime snapshot / change detection)
+    # ------------------------------------------------------------------
+    def regime_snapshot(self):
+        """Return a copy of the per-block contact regime codes.
+
+        Regime codes: ``-1`` = unset, ``0`` = interior (stick),
+        ``1`` = polar (separation), ``2`` = boundary (sliding).
+
+        Returns a 1-D ``ndarray`` (batch path) or ``dict`` (non-batch).
+        """
+        if self._uniform_m is not None:
+            return self._batch_branch_region.copy()
+        if self._branch_region:
+            return dict(self._branch_region)
+        return None
+
+    def regime_changed_mask(self, prev_snapshot, n_dof):
+        """Return a suppression mask for DOFs coupled to transitioning blocks.
+
+        Parameters
+        ----------
+        prev_snapshot : ndarray or dict
+            Value returned by a prior call to :meth:`regime_snapshot`.
+        n_dof : int
+            Total state vector dimension.
+
+        Returns
+        -------
+        mask : ndarray of float, shape ``(n_dof,)``
+            ``1.0`` for DOFs to keep, ``0.0`` for DOFs to suppress.
+            *None* if regime tracking is unavailable.
+
+        Notes
+        -----
+        Suppressed DOFs are determined by the ``_velocity_dof_map``
+        attribute (set externally, e.g. by :func:`build_impulse_contact`).
[…]
