+++ tests/test_active_set_filter.py
@@ -1,0 +1,461 @@
+"""Tests for the active-set error-norm filter.
+
+The filter suppresses velocity DOFs undergoing contact-regime transitions
+(stick↔slip, contact↔separation) from the adaptive error norm, preventing
+the embedded / Richardson error estimator from overreacting to the
+discontinuous constraint-force jumps intrinsic to nonsmooth event-capturing
+integrators.
+
+Test plan:
+  1. `regime_snapshot()` returns expected types and values
+  2. `regime_changed_mask()` produces correct masks
+  3. `CompositeContactProjection` delegates regime methods
+  4. `AdaptiveStepping._get_projection()` traverses solver chain
+  5. `build_impulse_contact` attaches `_velocity_dof_map`
+  6. End-to-end: filter active → larger accepted steps on a transition problem
+"""
+
+import numpy as np
+import pytest
+
+import solve_nivp
+from solve_nivp.projections import (
+    MuScaledSOCProjection,
+    CompositeContactProjection,
+    AlgebraicConstraintProjection,
+    Projection,
+    IdentityProjection,
+)
+from solve_nivp.contact import build_impulse_contact
+from solve_nivp.adaptive_integrator import AdaptiveStepping
+
+
+# =====================================================================
+# Helpers
+# =====================================================================
+
+def _make_soc_projection(nb=3, m=1, mu=0.5):
+    """Create a MuScaledSOCProjection with *nb* blocks of tang dim *m*."""
+    n_tot = nb * (1 + m)
+    blocks = []
+    for k in range(nb):
+        s_idx = k * (1 + m)
+        w_idx = list(range(s_idx + 1, s_idx + 1 + m))
+        blocks.append((s_idx, w_idx))
+    proj = MuScaledSOCProjection(
+        blocks=blocks,
+        get_mu=lambda y, _mu=mu: _mu,
+    )
+    return proj, n_tot
+
+
+def _bouncing_ball_contact_system(mu=0.3, e=0.0):
+    """Standard 2D bouncing ball returning ContactSystem."""
+    mass = 1.0
+    gravity = np.array([0.0, -9.81])
+    A = np.diag([mass, mass, 1.0, 1.0])
+
+    def rhs(t, y):
+        v = y[0:2]
+        return np.concatenate([mass * gravity, v])
+
+    def gap_func(y, t):
+        return np.array([y[3]])
+
+    y0 = np.array([2.0, 0.0, 0.0, 1.0])  # sliding + falling
+    contacts = [dict(vel_normal_idx=1, vel_tangential_idx=[0], mu=mu, e=e)]
+    cs = build_impulse_contact(
+        A, rhs, y0, contacts, gap_func,
+        theta=1.0,
+        component_slices=[slice(0, 2), slice(2, 4)],
+    )
+    return cs
+
+
+# =====================================================================
+# 1. regime_snapshot
+# =====================================================================
+
+class TestRegimeSnapshot:
+    """Test regime_snapshot on MuScaledSOCProjection (batch path)."""
+
+    def test_returns_ndarray(self):
+        proj, _ = _make_soc_projection(nb=4)
+        snap = proj.regime_snapshot()
+        assert isinstance(snap, np.ndarray)
+        assert snap.shape == (4,)
+        # All unset initially
+        assert np.all(snap == -1)
+
+    def test_snapshot_is_copy(self):
+        proj, _ = _make_soc_projection(nb=3)
+        snap1 = proj.regime_snapshot()
+        proj._batch_branch_region[0] = 2
+        snap2 = proj.regime_snapshot()
+        # snap1 should NOT reflect the change
+        assert snap1[0] == -1
+        assert snap2[0] == 2
+
[…]
