### 2026-03-09 #9775 @trintlermint
Good evening @danielskatz, I apologise for the extremely late review and why it is unnacceptable, and I apologise for my low contribution in this particular issue; however I was displaced due to the war efforts currently and had many travelling issues along with various other personal issues, which have now resolved and I have conducted full review in accordance with the T&Cs and pre-2026 guidelines, there are no more issues that arised in the newest version of the program after latest changes of the (3) issues raised by my fellow reviewers. Thank you and I apologise for my inability to inform you of this situation.
