+++ paper/biblio.bib
@@ -85,7 +85,7 @@
 
 @article{StewartTrinkle1996,
   author    = {D. E. Stewart and J. C. Trinkle},
-  title     = {An implicit time-stepping scheme for rigid body dynamics with inelastic collisions and Coulomb friction},
+  title     = {An implicit time-stepping scheme for rigid body dynamics with inelastic collisions and {C}oulomb friction},
   journal   = {International Journal for Numerical Methods in Engineering},
   year      = {1996},
   volume    = {39},
@@ -121,7 +121,7 @@
 
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -157,7 +157,7 @@ author = {Fabio Tramontana and Frank Westerhoff and Laura Gardini},
 }
 
 @article{qi1993nonsmooth,
-  title={A nonsmooth version of Newton's method},
+  title={A nonsmooth version of {N}ewton's method},
   author={Qi, Liqun and Sun, Jie},
   journal={Mathematical programming},
   volume={58},
@@ -169,13 +169,14 @@ author = {Fabio Tramontana and Frank Westerhoff and Laura Gardini},
 
 
 @article{armijo1966minimization,
-  title={Minimization of functions having Lipschitz continuous first partial derivatives},
+  title={Minimization of functions having {L}ipschitz continuous first partial derivatives},
   author={Armijo, Larry},
   journal={Pacific Journal of mathematics},
   volume={16},
   number={1},
   pages={1--3},
   year={1966},
+  doi={10.2140/pjm.1966.16.1},
   publisher={Mathematical Sciences Publishers}
 }
 
@@ -203,7 +204,7 @@ author = {Fabio Tramontana and Frank Westerhoff and Laura Gardini},
 
 @book{FortinGlowinski1983,
   author    = {Fortin, Michel and Glowinski, Roland},
-  title     = {Augmented Lagrangian Methods: Applications to the Numerical Solution of Boundary-Value Problems},
+  title     = {Augmented {L}agrangian Methods: Applications to the Numerical Solution of Boundary-Value Problems},
   publisher = {North-Holland},
   address   = {Amsterdam},
   year      = {1983}
@@ -224,7 +225,7 @@ doi = {10.1016/j.compstruc.2012.01.009},
 
 
 @book{hairer1993solving,
-  title={Solving Ordinary Differential Equations I: Nonstiff Problems},
+  title={Solving Ordinary Differential Equations {I}: Nonstiff Problems},
   author={Hairer, Ernst and N{\o}rsett, Syvert Paul and Wanner, Gerhard},
   year={1993},
   publisher={Springer},
