### 2026-03-09 #9775 @danielskatz
👋 @dmriley993 - I've suggested a few small changes in https://github.com/ERC-INJECT/solve_nivp/pull/4, including adding the DOI mentioned above - please merge this, or let me know what you disagree with, then we can move toward finishing the publication process
