### 2026-03-16 d1e1b63
chore: bump version to 0.1.1 in CITATION.cff
