+++ CITATION.cff
@@ -1,6 +1,6 @@
 cff-version: 1.2.0
 title: "solve_nivp: A Python toolkit for integrating nonsmooth dynamical systems"
-version: 0.1.0
+version: 0.1.1
 authors:
   - family-names: Riley
     given-names: David Michael
