### 2026-03-16 #9775 @dmriley993
Hi @danielskatz  we have created and archived the reviewed software release.

Version tag: v0.1.1
Archive DOI: 10.5281/zenodo.19046304
