### 2026-03-16 #9775 @danielskatz
Congratulations to @dmriley993 (David Riley) and co-author on your JOSS publication!!

And thanks to @ductho-le, @trintlermint, and @wwhenxuan for reviewing!
JOSS completely depends on volunteers - we couldn't do this without you
